requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
et-xmlfile==1.1.0
idna==3.4
mysql-connector-python==8.1.0
numpy==1.25.2
openpyxl==3.1.2
pandas==2.0.3
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.4
